Move the MyFile data class in a separate file
