+++ app/src/main/java/com/github/se/eduverse/model/folder/MyFile.kt
@@ -1,0 +1,11 @@
+package com.github.se.eduverse.model.folder
+
+import java.util.Calendar
+
+data class MyFile(
+    val name: String,
+    val creationTime: Calendar,
+    var lastAccess: Calendar,
+    var numberAccess: Int
+)
+
